Adds shared constants file and replaces old references
