+++ src/shared/constants.ts
@@ -1,0 +1,10 @@
+/**
+ *  Constants required by both the background and popup components.
+ */
+
+export const popupDimensions = {
+  defaultWidth: 300,
+  defaultHeight: 480,
+  expandedWidth: 430,
+  expandedHeight: 820
+};
